+++ src/crypto/salsa2012.rs
@@ -14,5 +14,5 @@ use randombytes::randombytes_into;
 
 stream_module!(crypto_stream_salsa2012,
                crypto_stream_salsa2012_xor,
-               crypto_stream_salsa2012_KEYBYTES as usize,
-               crypto_stream_salsa2012_NONCEBYTES as usize);
+               crypto_stream_salsa2012_KEYBYTES,
+               crypto_stream_salsa2012_NONCEBYTES);
